refactor: rename accessors→queries, split session tests, remove unnecessary clone
- Rename src/app/accessors.rs → src/app/queries.rs (read-only query
  methods on TuiMode; aligns with sibling module naming convention)
- Split session/mod.rs (533→314 lines) per ADR-044 ~300-line ceiling:
  extract compact.rs (111 lines, 4 tests) and fork.rs (113 lines, 5 tests)
- Remove unnecessary .clone() in run_proactive_compaction: reborrow
  from &mut self to &self is sufficient for should_compact_proactively
- Update CONTRIBUTING.md and REPO-RAW-URL-MAP.md references
