+++ src/state/conversation/history.rs
@@ -238,8 +238,7 @@ impl ConversationManager {
         &mut self,
         context_window_tokens: usize,
     ) -> Option<(usize, usize, String)> {
-        let config = self.compaction_config.clone();
-        if !self.should_compact_proactively(&config, context_window_tokens) {
+        if !self.should_compact_proactively(&self.compaction_config, context_window_tokens) {
             return None;
         }
         let messages_before = self.api_messages.len();
